Crate service

metw-accounts-center service

This crate is the abstract orchestrating microservice for metw.cc. It manages clients and data repositories, and provides a library interface to use with a presentation layer.

System Design

This project strictly adheres to the Separation of Concerns (SoC) principle. The service abstraction layers are:

repo: Defines traits for application storage, database, and cache access.

service: Provides an interface for manipulating application state. Elaborates on low-level transactions defined in the repo layer to the upper layers.

handlers: The final level of abstraction to manipulate the application state. This layer is a transport-agnostic front end for the microservice.

AppState: The application state contains the services constructed from repositories and clients. Handlers require a state instance to construct.

Supporting Modules

client is used for external integrations (e.g., email provider, CAPTCHA).

The dto module defines the interface to communicate with handlers. dto::service, dto::repo and entity are used for internal communication between service and repo layers.

util is used for miscellaneous utilities that do not fit into categories above. Check out documentations of the util for details. Conceptually, the token and id modules are re-exported at the crate root as they are too common.

Re-exports

pub use util::id;

pub use util::token;

Modules

client

External integrations.

dto

Data transfer objects.

entity

Database entities.

handlers

An interface for manipulating the application state.

repo

The data access layer.

service

An interface for manipulating a repository.

testutiltestutil

Test utilities.

util

Miscellaneous utilities.

Structs

AppState

Application-wide state.

Functions

checked_now

Current timestamp, with clock skew detection.