service::handlers

Struct SessionHandler

pub struct SessionHandler(pub AppState);

Account handlers that does not require escalated privileges.

This handlers should be protected using tokens with TokenScope::Session, account_id parameters in methods of this struct extracted from that token.

Tuple Fields
0: AppState
Implementations

impl SessionHandler

pub async fn get_current_account( self, account_id: AccountId, ) -> HandlerResult<Account>

Returns account details.

pub async fn add_email( self, account_id: AccountId, email_dto: Email, ip: IpAddr, captcha: Captcha, ) -> HandlerResult<()>

Sends ConfirmNewEmail to add requested email.

pub async fn remove_email( self, account_id: AccountId, email_dto: Email, ) -> HandlerResult<()>

Removes the email if it is not account’s primary email.

pub async fn change_primary_email( self, account_id: AccountId, email_dto: Email, captcha: Captcha, ) -> HandlerResult<()>

Sends ConfirmPrimaryEmailChange email to current primary email.

pub async fn change_password( &self, account_id: AccountId, change_password_dto: ChangePassword, ) -> HandlerResult<()>

Update account password.

pub async fn get_owned_applications( &self, account_id: AccountId, ) -> HandlerResult<Vec<ApplicationSummary>>

Gets list of the applications owned by the account.

pub async fn create_application( &self, account_id: AccountId, create_app_dto: ApplicationName, captcha: Captcha, ) -> HandlerResult<CreatedApplication>

Register a new application.

pub async fn get_application_consents( &self, account_id: AccountId, pagination_dto: ConsentPagination, ) -> HandlerResult<Vec<ApplicationConsent>>

Gets list of the applications owned by the account.

pub async fn authorize_application( &self, account_id: AccountId, application_id: ApplicationId, ) -> HandlerResult<()>

Authorize the application.

pub async fn unauthorize_application( &self, account_id: AccountId, application_id: ApplicationId, ) -> HandlerResult<()>

Remove authorization of the application.

pub async fn get_application_consent_status( &self, account_id: AccountId, application_id: ApplicationId, ) -> HandlerResult<ApplicationConsentStatus>

Check the authorization of the application.