service::repo

Trait ApplicationRepo

pub trait ApplicationRepo: Send + Sync {
    // Required methods
    fn begin_transaction<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = RepoResult<Box<dyn ApplicationRepoTransaction>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn list_by_owner<'life0, 'async_trait>(
        &'life0 self,
        account_id: AccountId,
    ) -> Pin<Box<dyn Future<Output = RepoResult<Vec<OwnedApplicationSummary>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn list_redirect_urls<'life0, 'async_trait>(
        &'life0 self,
        application_id: ApplicationId,
    ) -> Pin<Box<dyn Future<Output = RepoResult<Vec<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn is_owned_by<'life0, 'async_trait>(
        &'life0 self,
        application_id: ApplicationId,
        account_id: AccountId,
    ) -> Pin<Box<dyn Future<Output = RepoResult<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

3rd party applications - user-registered authorization consent screen.

Required Methods

fn begin_transaction<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = RepoResult<Box<dyn ApplicationRepoTransaction>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Begin a new transactional unit.

fn list_by_owner<'life0, 'async_trait>( &'life0 self, account_id: AccountId, ) -> Pin<Box<dyn Future<Output = RepoResult<Vec<OwnedApplicationSummary>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets list of the applications owned by the account.

fn list_redirect_urls<'life0, 'async_trait>( &'life0 self, application_id: ApplicationId, ) -> Pin<Box<dyn Future<Output = RepoResult<Vec<String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets list of the redirect URLs of the application.

fn is_owned_by<'life0, 'async_trait>( &'life0 self, application_id: ApplicationId, account_id: AccountId, ) -> Pin<Box<dyn Future<Output = RepoResult<bool>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Whether or not the application is owned by given account.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".