service::repo

Trait AccountRepo

pub trait AccountRepo: Send + Sync {

    // Required methods
    fn begin_transaction<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = RepoResult<Box<dyn AccountRepoTransaction>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_login_by_email<'life0, 'life1, 'async_trait>(
        &'life0 self,
        email: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = RepoResult<Option<Login>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_login_by_username<'life0, 'life1, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = RepoResult<Option<Login>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_primary_username<'life0, 'async_trait>(
        &'life0 self,
        id: AccountId,
    ) -> Pin<Box<dyn Future<Output = RepoResult<Option<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_nonexpiring_username_aliases<'life0, 'async_trait>(
        &'life0 self,
        id: AccountId,
    ) -> Pin<Box<dyn Future<Output = RepoResult<Vec<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_primary_email<'life0, 'async_trait>(
        &'life0 self,
        id: AccountId,
    ) -> Pin<Box<dyn Future<Output = RepoResult<Option<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_secondary_emails<'life0, 'async_trait>(
        &'life0 self,
        id: AccountId,
    ) -> Pin<Box<dyn Future<Output = RepoResult<Vec<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_keys<'life0, 'async_trait>(
        &'life0 self,
        id: AccountId,
    ) -> Pin<Box<dyn Future<Output = RepoResult<Option<Keys>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_account_flags<'life0, 'async_trait>(
        &'life0 self,
        id: AccountId,
    ) -> Pin<Box<dyn Future<Output = RepoResult<Option<AccountFlags>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_primary_email_if_current_is<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        id: AccountId,
        current_primary_email: &'life1 str,
        new_primary_email: &'life2 str,
    ) -> Pin<Box<dyn Future<Output = RepoResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn remove_email_if_not_primary<'life0, 'life1, 'async_trait>(
        &'life0 self,
        id: AccountId,
        email: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = RepoResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn is_username_taken<'life0, 'life1, 'async_trait>(
        &'life0 self,
        username: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = RepoResult<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn is_email_taken<'life0, 'life1, 'async_trait>(
        &'life0 self,
        email: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = RepoResult<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn is_email_taken_by<'life0, 'life1, 'async_trait>(
        &'life0 self,
        id: AccountId,
        email: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = RepoResult<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;

}

Persistent account storage.

Required Methods

fn begin_transaction<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = RepoResult<Box<dyn AccountRepoTransaction>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Begin a new transactional unit.

fn get_login_by_email<'life0, 'life1, 'async_trait>( &'life0 self, email: &'life1 str, ) -> Pin<Box<dyn Future<Output = RepoResult<Option<Login>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get password by email.

Returns Argon2-hashed password with account id. Higher layers shall do hash verifications.

fn get_login_by_username<'life0, 'life1, 'async_trait>( &'life0 self, username: &'life1 str, ) -> Pin<Box<dyn Future<Output = RepoResult<Option<Login>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get password by username.

The behavior is exactly the same with get_login_by_email.

fn get_primary_username<'life0, 'async_trait>( &'life0 self, id: AccountId, ) -> Pin<Box<dyn Future<Output = RepoResult<Option<String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get primary username if, exists.

Usually all the accounts have primary usernames, but pending deletion might drop usernames of an account.

fn get_nonexpiring_username_aliases<'life0, 'async_trait>( &'life0 self, id: AccountId, ) -> Pin<Box<dyn Future<Output = RepoResult<Vec<String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get usernames by account id.

After username change, previous usernames for will continue to belong the account for a while; but they will be garbage collected.

fn get_primary_email<'life0, 'async_trait>( &'life0 self, id: AccountId, ) -> Pin<Box<dyn Future<Output = RepoResult<Option<String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get primary email if, exists.

All regular accounts shall have primary mail, but some system accounts or deleted accounts do not.

fn get_secondary_emails<'life0, 'async_trait>( &'life0 self, id: AccountId, ) -> Pin<Box<dyn Future<Output = RepoResult<Vec<String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get secondary emails by account id.

Returns an empty list if none have been added.

fn get_keys<'life0, 'async_trait>( &'life0 self, id: AccountId, ) -> Pin<Box<dyn Future<Output = RepoResult<Option<Keys>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get account keys - the key bundle of the account.

fn get_account_flags<'life0, 'async_trait>( &'life0 self, id: AccountId, ) -> Pin<Box<dyn Future<Output = RepoResult<Option<AccountFlags>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get account flags.

fn set_primary_email_if_current_is<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, id: AccountId, current_primary_email: &'life1 str, new_primary_email: &'life2 str, ) -> Pin<Box<dyn Future<Output = RepoResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Set the email primary for the account.

Although emails are unique, the id parameter is also required to prevent race conditions. It is highly unlikely that the owner of an email would change at the exact moment their email is being set as the primary email, but it is still a safety hazard that must still be prevented.

fn remove_email_if_not_primary<'life0, 'life1, 'async_trait>( &'life0 self, id: AccountId, email: &'life1 str, ) -> Pin<Box<dyn Future<Output = RepoResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Remove the email if it is not primary mail of the account.

fn is_username_taken<'life0, 'life1, 'async_trait>( &'life0 self, username: &'life1 str, ) -> Pin<Box<dyn Future<Output = RepoResult<bool>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Returns true if the username has been taken.

fn is_email_taken<'life0, 'life1, 'async_trait>( &'life0 self, email: &'life1 str, ) -> Pin<Box<dyn Future<Output = RepoResult<bool>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Returns true if the email has been taken.

fn is_email_taken_by<'life0, 'life1, 'async_trait>( &'life0 self, id: AccountId, email: &'life1 str, ) -> Pin<Box<dyn Future<Output = RepoResult<bool>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Returns true if the email has been taken by the given account.